handle +99 in the exponent.  Integer.parseInt("+99") fails, so   use ParseDouble instead.
svn path=/dasCore/trunk/; revision=1832

+++ dasCore/src/edu/uiowa/physics/pw/das/datum/NumberUnits.java
@@ -96,7 +96,7 @@ public class NumberUnits extends Units {
             } else {
                 mant= s.substring(0,ie);
                 double[] dd= parseDecimal( mant );
-                int exp= Integer.parseInt( s.substring(ie+1) );
+                int exp= (int)Double.parseDouble( s.substring(ie+1) );
                 dd[0]= dd[0] * DasMath.exp10(exp);
                 dd[1]= dd[1] * DasMath.exp10(exp);
                 return dd;
